fixed casting bug

+++ src/main/java/com/c123/MySQLConsole/dao/UserDAO.java
@@ -8,7 +8,7 @@ public interface UserDAO {
 
     List<User> getall();
     User getOne(String envId);
-    Integer getConnections(String envId);
+    Long getConnections(String envId);
 
     User create(String envId, String host, String password);
